requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1

# Security dependencies
Flask-Talisman==1.1.0  # Security headers and HTTPS enforcement
cryptography==41.0.7   # Secure cryptographic operations

# Web scraping and parsing
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Skill extraction dependencies
spacy==3.8.7
skillner==1.0.3

# Development dependencies
ipython==8.18.1

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
coverage==7.3.2

# Performance and monitoring
Flask-Caching==2.1.0   # Caching support
redis==5.0.1           # Redis for caching (optional)

# Database migrations (optional but recommended)
Flask-Migrate==4.0.5
